# src/message/controllers/messageController.ts
import type { Logger } from '@map-colonies/js-logger';
import httpStatus from 'http-status-codes';
import { injectable, inject } from 'tsyringe';
import { type Registry, Counter } from 'prom-client';
import type { TypedRequestHandlers } from '@openapi';
import { SERVICES } from '@common/constants';
import { MessageManager, ILogObject } from '../models/messageManager';
import { localMessagesStore } from '../../common/mocks';
import { IQueryModel } from './../../common/interfaces';

@injectable()
export class MessageController {
  private readonly createdMessageCounter: Counter;

  public constructor(
    @inject(SERVICES.LOGGER) private readonly logger: Logger,
    @inject(MessageManager) private readonly manager: MessageManager,
    @inject(SERVICES.METRICS) private readonly metricsRegistry: Registry
  ) {
    this.createdMessageCounter = new Counter({
      name: 'created_message',
      help: 'number of created message',
      registers: [this.metricsRegistry],
    });
  }

  public createMessage: TypedRequestHandlers['POST /message'] = (req, res) => {
    try {
      const newMessage = this.manager.createMessage(req.body);

      this.createdMessageCounter.inc(1);
      return res.status(httpStatus.CREATED).json({ id: newMessage.id });
    } catch (error) {
      this.logger.error({ msg: 'Error creating message', error });
      return res.status(httpStatus.INTERNAL_SERVER_ERROR).json({ message: 'Failed to create message' });
    }
  };

  public getMessages: TypedRequestHandlers['GET /message'] = (req, res) => {
    try {
      const params: IQueryModel | undefined = req.query;

      const hasParams = !!params && Object.keys(params).length > 0;

      const resultMessages: ILogObject[] = hasParams ? this.manager.getMessages(params) : localMessagesStore;

      return res.status(httpStatus.OK).json(resultMessages);
    } catch (error) {
      this.logger.error({ msg: 'Error retrieving messages', error });
      return res.status(httpStatus.INTERNAL_SERVER_ERROR).json({ message: 'Failed to get messages' });
    }
  };

  public getMessageById: TypedRequestHandlers['GET /message/{id}'] = (req, res) => {
    try {
      const id = req.params.id;

      const message = this.manager.getMessageById(id);

      if (!message) {
        return res.status(httpStatus.NOT_FOUND).json({ message: `No message found with id '${id}'` });
      }

      return res.status(httpStatus.OK).json(message);
    } catch (error) {
      this.logger.error({ msg: 'Error retrieving message', error });
      return res.status(httpStatus.INTERNAL_SERVER_ERROR).json({ message: 'Failed to get message by id' });
    }
  };

  public deleteMessageById: TypedRequestHandlers['DELETE /message/{id}'] = (req, res) => {
    try {
      const id = req.params.id;

      const message = this.manager.deleteMessageById(id);

      if (!message) {
        return res.status(httpStatus.NOT_FOUND).json({ message: `No message found to delete with id '${id}'` });
      }

      return res.status(httpStatus.OK).json({ id: `Message with id '${id}' was deleted successfully` });
    } catch (error) {
      this.logger.error({ msg: 'Error retrieving message', error });
      return res.status(httpStatus.INTERNAL_SERVER_ERROR).json({ message: 'Failed to delete message' });
    }
  };
}